// src/components/models/molecules/ProductsCards.ts
export interface ProductInternalProps {
    idProducto: number;
    codigo: string;
    nombre: string;
    descripcion: string;
    categoria: string;
    marca: string;
    imagen: string;
    isActive: string;
    stock: number;
    precio_compra: number;
    precio_venta: number;
    inversion: number;
}